requirements.txt
fastapi
uvicorn[standard]
sqlmodel
httpx
python-dotenv
pydantic
alembic
